FIX: comment validation,dialog section
